springbootandjpa/src/main/java/com/yohwan/springbootandjpa/service/MemberService.java
package com.yohwan.springbootandjpa.service;

import java.util.List;

import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import com.yohwan.springbootandjpa.domain.Member;
import com.yohwan.springbootandjpa.repository.MemberRepository;

import lombok.RequiredArgsConstructor;

@Service
@RequiredArgsConstructor
public class MemberService {
    
    private final MemberRepository memberRepository;

    @Transactional
    public Long join(Member member){
        validateDuplicateMember(member);
        return memberRepository.save(member).getId();

    }

    @Transactional(readOnly = true)
    private void validateDuplicateMember(Member member) {
        List<Member> findMembers = memberRepository.findByName(member.getName()); //멤버네임꼭유니크제약
        if(!findMembers.isEmpty()){
            throw new IllegalStateException("이미존재하는회원");
        }
    }

    @Transactional(readOnly = true)
    public List<Member> findMembers() {
        return memberRepository.findAll();
    }

    @Transactional(readOnly = true)
    public Member findOne(Long memberId){
        return memberRepository.findById(memberId).get();
    }

    @Transactional
    public void update(Long id, String name) {
        Member member = memberRepository.findById(id).get();
        member.setName(name);
    }
}
