springbootandjpa/src/test/java/com/yohwan/springbootandjpa/service/MemberServiceTest.java
package com.yohwan.springbootandjpa.service;

import static org.junit.Assert.*;

import org.junit.Assert;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.TestPropertySource;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.transaction.annotation.Transactional;

import com.yohwan.springbootandjpa.domain.Member;
import com.yohwan.springbootandjpa.repository.MemberRepositoryOld;


@RunWith(SpringRunner.class)
@SpringBootTest
@Transactional
@TestPropertySource(properties = "spring.config.location=classpath:application-test.yml" )
public class MemberServiceTest {

    @Autowired MemberService memberService;
    @Autowired MemberRepositoryOld memberRepository;

    @Test
    public void 회원가입() {

        //given
        Member member = new Member();
        member.setName("kim");

        //when
        Long savedId = memberService.join(member);

        //then
        Assert.assertEquals(member, memberRepository.findOne(savedId));
    }

    @Test(expected = IllegalStateException.class)
    public void 중복회원예외() {

        //given
        Member member1 = new Member();
        member1.setName("kim1");

        Member member2 = new Member();
        member2.setName("kim1");

        //when
        memberService.join(member1);
        memberService.join(member2);


        //then
        fail("예외발생");

    }

}
